IPFIX: switch from generated encoder to parameter driven encoding
Generating templates from datasets makes it difficult
(impossible) to follow the template withdrawel/reuse
rules. We need to be able to define template before
generating datasets.

To be  abel to generate template records without having to
lock a the data sets, the type and size of each attribute
has to be known. The same information can be used for
parameter driven value encoding (instead of the generated
encoder we had before).

This switches to a generated information element to type mapping.

The next change will use the new mapping to generate templates
without the need to for real datasets.